add precache command that can cause IGV to preload data into memory
This allows background applications to save the user time by loading
the regions of interest into memory before the user goes there. This
can save significant time if the user is navigating through many
regions over a slow connection. The regions are pinned in memory,
so it is important that the user will clear the session to free
this memory, otherwise eventually memory will run out.

Example link to precache region:

http://localhost:60151/execute?command=precache+chr3:38,546,062-38,651,687+chr7:117,478,025-117,670,665'